+++ openapi_schema_pydantic/v3/v3_0_3/encoding.py
@@ -60,7 +60,7 @@ class Encoding(BaseModel):
     """
 
     class Config:
-        extra = Extra.forbid
+        extra = Extra.ignore
         schema_extra = {
             "examples": [
                 {