+++ openapi_schema_pydantic/v3/v3_0_3/link.py
@@ -63,7 +63,7 @@ class Link(BaseModel):
     """
 
     class Config:
-        extra = Extra.forbid
+        extra = Extra.ignore
         schema_extra = {
             "examples": [
                 {"operationId": "getUserAddressByUUID", "parameters": {"userUuid": "$response.body#/uuid"}},