+++ openapi_schema_pydantic/v3/v3_0_3/reference.py
@@ -16,7 +16,7 @@ class Reference(BaseModel):
     """**REQUIRED**. The reference string."""
 
     class Config:
-        extra = Extra.forbid
+        extra = Extra.ignore
         allow_population_by_field_name = True
         schema_extra = {
             "examples": [{"$ref": "#/components/schemas/Pet"}, {"$ref": "Pet.json"}, {"$ref": "definitions.json#/Pet"}]