+++ openapi_schema_pydantic/v3/v3_0_3/security_scheme.py
@@ -66,7 +66,7 @@ class SecurityScheme(BaseModel):
     """
 
     class Config:
-        extra = Extra.forbid
+        extra = Extra.ignore
         allow_population_by_field_name = True
         schema_extra = {
             "examples": [