+++ openapi_schema_pydantic/v3/v3_0_3/xml.py
@@ -48,7 +48,7 @@ class XML(BaseModel):
     """
 
     class Config:
-        extra = Extra.forbid
+        extra = Extra.ignore
         schema_extra = {
             "examples": [
                 {"namespace": "http://example.com/schema/sample", "prefix": "sample"},