+++ openapi_schema_pydantic/v3/v3_1_0/operation.py
@@ -107,7 +107,7 @@ class Operation(BaseModel):
     """
 
     class Config:
-        extra = Extra.forbid
+        extra = Extra.ignore
         schema_extra = {
             "examples": [
                 {