+++ openapi_schema_pydantic/v3/v3_1_0/parameter.py
@@ -141,7 +141,7 @@ class Parameter(BaseModel):
     """
 
     class Config:
-        extra = Extra.forbid
+        extra = Extra.ignore
         allow_population_by_field_name = True
         schema_extra = {
             "examples": [